polymarket-weather/assets/polymarket_weather_raw/requirements.txt
pandas
requests
meteostat
google-cloud-bigquery
db-dtypes
pyyaml
